refactor:  useTeams Hook 역할 분리
