+++ frontend/src/apis/utils.ts
@@ -4,12 +4,7 @@ import { ShowSnackbarProps } from 'hooks/utils/useSnackbar';
 
 import { MESSAGE } from 'constants/constants';
 
-export const 엑세스토큰이없는경우헤더제거 = ({
-  accessToken,
-  method,
-  url,
-  headers,
-}: RequestType) => {
+export const NotAccessTokenRemoveHeader = ({ accessToken, method, url, headers }: RequestType) => {
   if (!accessToken) {
     return {
       accessToken,
