+++ frontend/src/components/teams/Teams.tsx
@@ -10,18 +10,14 @@ import InterviewTeam from 'components/teams/InterviewTeam';
 import { InterviewTeamType, TeamConditionsType } from 'types/team';
 
 const Teams = ({ teamsCondition }: TeamsProps) => {
-  const { getTeams } = useTeams();
-
-  const getTeamsQuery = useQuery(['teams', teamsCondition], () => getTeams({ teamsCondition }), {
-    suspense: true,
-  });
+  const { teams } = useTeams(teamsCondition);
 
   return (
     <S.Container>
-      {getTeamsQuery.data?.teams.map((team: InterviewTeamType) => (
+      {teams?.map((team: InterviewTeamType) => (
         <InterviewTeam key={team.id} team={team} />
       ))}
-      {getTeamsQuery.data?.teams.length === 0 && <EmptyTeams />}
+      {teams?.length === 0 && <EmptyTeams />}
     </S.Container>
   );
 };
