+++ frontend/src/hooks/team/useTeams.ts
@@ -1,6 +1,6 @@
-import { useState } from 'react';
-
 import axios, { AxiosResponse } from 'axios';
+
+import { useQuery } from '@tanstack/react-query';
 
 import useSnackbar from 'hooks/utils/useSnackbar';
 
@@ -9,39 +9,28 @@ import { TEAMS_CONDITION } from 'constants/constants';
 import { requestGetMyTeams } from 'apis/myInfo';
 import { requestGetTeams } from 'apis/teams';
 import { NotCorrectToken } from 'apis/utils';
-import { InterviewTeamType, TeamApiType, TeamConditionsType } from 'types/team';
+import { TeamConditionsType } from 'types/team';
 
-const useTeams = () => {
+const useTeams = (teamsCondition: TeamConditionsType) => {
   const { showSnackbar } = useSnackbar();
-
-  const [teams, setTeams] = useState<InterviewTeamType[]>([]);
-  const [teamsCondition, setTeamsCondition] = useState({
-    open: true,
-    close: false,
-    my: false,
-  });
-
   const accessToken = localStorage.getItem('accessToken');
 
   const getTeams = async ({ teamsCondition }: Record<'teamsCondition', TeamConditionsType>) => {
     try {
       if (teamsCondition.open) {
         const res = await requestGetTeams({ accessToken, teamsCondition: TEAMS_CONDITION.OPEN });
-        setTeams(res.data.teams);
 
         return res.data;
       }
 
       if (teamsCondition.close) {
         const res = await requestGetTeams({ accessToken, teamsCondition: TEAMS_CONDITION.CLOSE });
-        setTeams(res.data.teams);
 
         return res.data;
       }
 
       if (teamsCondition.my) {
         const res = await requestGetMyTeams({ accessToken });
-        setTeams(res.data.teams);
 
         return res.data;
       }
@@ -55,33 +44,12 @@ const useTeams = () => {
     }
   };
 
-  const handleClickFilterButtons = (e: React.MouseEvent<HTMLElement>) => {
-    const eventTarget = e.target as HTMLElement;
-
-    switch (eventTarget.innerText) {
-      case '진행중인 인터뷰':
-        if (teamsCondition.open) return;
-
-        setTeamsCondition({ open: true, close: false, my: false });
-        break;
-      case '종료된 인터뷰':
-        if (teamsCondition.close === true) return;
-
-        setTeamsCondition({ open: false, close: true, my: false });
-        break;
-      case '나의 인터뷰':
-        if (teamsCondition.my) return;
-
-        setTeamsCondition({ open: false, close: false, my: true });
-        break;
-    }
-  };
+  const { data: teams } = useQuery(['teams', teamsCondition], () => getTeams({ teamsCondition }), {
+    suspense: true,
+  });
 
   return {
-    teams,
-    getTeams,
-    teamsCondition,
-    handleClickFilterButtons,
+    teams: teams?.teams,
   };
 };
